Refactor nosignal module: move to video package and rename VideoInfo to NsVideoInfo
